Prevent default action after handling keyboard shortcut
We fixed the problem for the default browser keybindings by checking for
all modifiers, but there's still the possibility that the user's browser
has some custom keybindings that conflict with ours. In this situation it
would probably be undesirable for the browser to also take its action, so
tell it that we've already handled the situation.